+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(c,i)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let t={};const r=e=>a(e,n),f={module:{uri:n},exports:t,require:r};s[n]=Promise.all(c.map((e=>f[e]||r(e)))).then((e=>(i(...e),t)))}}define(["./workbox-2780d724"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/106-92d741f60a89224c.js",revision:"92d741f60a89224c"},{url:"/_next/static/chunks/178-f053af60ba90a84f.js",revision:"f053af60ba90a84f"},{url:"/_next/static/chunks/1a48c3c1-f1a1a0cd4dc19b45.js",revision:"f1a1a0cd4dc19b45"},{url:"/_next/static/chunks/202-b8608992bfad2e62.js",revision:"b8608992bfad2e62"},{url:"/_next/static/chunks/258-49f2507e07c723cd.js",revision:"49f2507e07c723cd"},{url:"/_next/static/chunks/490-53786846faa8a941.js",revision:"53786846faa8a941"},{url:"/_next/static/chunks/536-244535e045a076e2.js",revision:"244535e045a076e2"},{url:"/_next/static/chunks/542.aff58bcde21fcebe.js",revision:"aff58bcde21fcebe"},{url:"/_next/static/chunks/61-143f7360a9128199.js",revision:"143f7360a9128199"},{url:"/_next/static/chunks/821-f53e4aad7a01a4bc.js",revision:"f53e4aad7a01a4bc"},{url:"/_next/static/chunks/894.884bc9854a9cfe85.js",revision:"884bc9854a9cfe85"},{url:"/_next/static/chunks/910-efedfa2b79469cd5.js",revision:"efedfa2b79469cd5"},{url:"/_next/static/chunks/991-0cc021cc659a70e5.js",revision:"0cc021cc659a70e5"},{url:"/_next/static/chunks/framework-a87821de553db91d.js",revision:"a87821de553db91d"},{url:"/_next/static/chunks/main-21fbc9a5ad2df4ee.js",revision:"21fbc9a5ad2df4ee"},{url:"/_next/static/chunks/pages/404-807af44f13f530bb.js",revision:"807af44f13f530bb"},{url:"/_next/static/chunks/pages/_app-f107f069aec897e7.js",revision:"f107f069aec897e7"},{url:"/_next/static/chunks/pages/_error-815e492bede44f3e.js",revision:"815e492bede44f3e"},{url:"/_next/static/chunks/pages/cartoons-63769e0e4ce8e9b9.js",revision:"63769e0e4ce8e9b9"},{url:"/_next/static/chunks/pages/favourites-19a0c81c8b3442a1.js",revision:"19a0c81c8b3442a1"},{url:"/_next/static/chunks/pages/film/%5Bid%5D-45a22eadb831ff2d.js",revision:"45a22eadb831ff2d"},{url:"/_next/static/chunks/pages/films-e9b747c9e3a115ff.js",revision:"e9b747c9e3a115ff"},{url:"/_next/static/chunks/pages/index-2e783af8fc6dee2f.js",revision:"2e783af8fc6dee2f"},{url:"/_next/static/chunks/pages/login-8a4fec9939274311.js",revision:"8a4fec9939274311"},{url:"/_next/static/chunks/pages/name/%5Bid%5D-56999971128384c0.js",revision:"56999971128384c0"},{url:"/_next/static/chunks/pages/register-dc48a39ad214624d.js",revision:"dc48a39ad214624d"},{url:"/_next/static/chunks/pages/room/%5Bid%5D-151c26a4e64dc4e9.js",revision:"151c26a4e64dc4e9"},{url:"/_next/static/chunks/pages/search/%5Bid%5D-0635ee16f9e0a6fa.js",revision:"0635ee16f9e0a6fa"},{url:"/_next/static/chunks/pages/series-69da5cdcbac5b1f4.js",revision:"69da5cdcbac5b1f4"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-08d05dc8a1f47353.js",revision:"08d05dc8a1f47353"},{url:"/_next/static/css/0c24d2c5563bb4d1.css",revision:"0c24d2c5563bb4d1"},{url:"/_next/static/css/15e4c5310504cd54.css",revision:"15e4c5310504cd54"},{url:"/_next/static/css/22d5d7579d114f75.css",revision:"22d5d7579d114f75"},{url:"/_next/static/css/390579dc1b8fbe64.css",revision:"390579dc1b8fbe64"},{url:"/_next/static/css/62daf9081586b583.css",revision:"62daf9081586b583"},{url:"/_next/static/css/81e102bbba6f5df3.css",revision:"81e102bbba6f5df3"},{url:"/_next/static/css/ac1370f24fa37b0c.css",revision:"ac1370f24fa37b0c"},{url:"/_next/static/css/b4bf9abad4c08e27.css",revision:"b4bf9abad4c08e27"},{url:"/_next/static/css/bda97fb7cc7e3a91.css",revision:"bda97fb7cc7e3a91"},{url:"/_next/static/css/bec5ff6efd5abea0.css",revision:"bec5ff6efd5abea0"},{url:"/_next/static/css/ea0a40089e461740.css",revision:"ea0a40089e461740"},{url:"/_next/static/media/lg.955a4bcf.woff2",revision:"955a4bcf"},{url:"/_next/static/media/lg.a5ca0178.svg",revision:"a5ca0178"},{url:"/_next/static/media/lg.c950f0b5.woff",revision:"c950f0b5"},{url:"/_next/static/media/lg.dc565ab5.ttf",revision:"dc565ab5"},{url:"/_next/static/media/loading.49ca460c.gif",revision:"49ca460c"},{url:"/_next/static/nZMb-eWgB03I6AF6ri6G3/_buildManifest.js",revision:"15a911a369818fadb282403b2185349b"},{url:"/_next/static/nZMb-eWgB03I6AF6ri6G3/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/nZMb-eWgB03I6AF6ri6G3/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/bazon.txt",revision:"f791ddcacca63ba73b03e35834720348"},{url:"/bg1.jpg",revision:"d75fc0ac993186a6d2f772fb906eabeb"},{url:"/favicon.ico",revision:"0b5d739697aa36f8f02fcc05cf17b2bd"},{url:"/fonts/FiraSans-Bold.woff2",revision:"07b9daf36cfedbb07b47c1eed9921763"},{url:"/fonts/FiraSans-ExtraBold.woff2",revision:"84e01280cf056a4198d6c14b2f8db407"},{url:"/fonts/FiraSans-Light.woff2",revision:"c2ece2ce8559a936160dd98c0747c9b2"},{url:"/fonts/FiraSans-Medium.woff2",revision:"3392354b6a7a83ac879b2dd4e258e7e1"},{url:"/fonts/FiraSans-Regular.woff2",revision:"44af09a058cd81661fc93376ee5d8f34"},{url:"/fonts/FiraSans-SemiBold.woff2",revision:"0b6a215f82eda2e94a2734e24a3924df"},{url:"/icons/icon-128x128.png",revision:"56d4125eeb03dcd0ec45a8b7c8448123"},{url:"/icons/icon-144x144.png",revision:"c616d2acddf9ad9c887616fee28998d9"},{url:"/icons/icon-152x152.png",revision:"4166a839dff25fa67fbc91461cc0d0f9"},{url:"/icons/icon-192x192.png",revision:"9dd9391d29f0a94f6a5014952b5bc94a"},{url:"/icons/icon-384x384.png",revision:"14e26ef81ef4b67665ea9af23e6fadcb"},{url:"/icons/icon-512x512.png",revision:"6302c46f65bb9ae7d18298a8e9815039"},{url:"/icons/icon-72x72.png",revision:"99f9d2024ae87162655c5120fc41cc74"},{url:"/icons/icon-96x96.png",revision:"689d68e5664c97d5bacb763a8c3468a3"},{url:"/kinobd.txt",revision:"b6d00bafd9a20676f430e8f3adbf3f41"},{url:"/logo.svg",revision:"aa4cf4f36c21cbbf39d36c4ef9455f78"},{url:"/manifest.json",revision:"127550a4d147e1e46235023b66936c18"},{url:"/player.js",revision:"ad82e20958c8ebd61c0e030040a18748"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(c,i)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let t={};const r=e=>a(e,n),f={module:{uri:n},exports:t,require:r};s[n]=Promise.all(c.map((e=>f[e]||r(e)))).then((e=>(i(...e),t)))}}define(["./workbox-2780d724"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/I4MBeZ7qj72mau7RZY2j5/_buildManifest.js",revision:"fec823b8c98159933091965fc6a052d9"},{url:"/_next/static/I4MBeZ7qj72mau7RZY2j5/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/I4MBeZ7qj72mau7RZY2j5/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/106-9a504ab08ba08203.js",revision:"9a504ab08ba08203"},{url:"/_next/static/chunks/178-f053af60ba90a84f.js",revision:"f053af60ba90a84f"},{url:"/_next/static/chunks/1a48c3c1-f1a1a0cd4dc19b45.js",revision:"f1a1a0cd4dc19b45"},{url:"/_next/static/chunks/202-b8608992bfad2e62.js",revision:"b8608992bfad2e62"},{url:"/_next/static/chunks/258-49f2507e07c723cd.js",revision:"49f2507e07c723cd"},{url:"/_next/static/chunks/490-53786846faa8a941.js",revision:"53786846faa8a941"},{url:"/_next/static/chunks/536-244535e045a076e2.js",revision:"244535e045a076e2"},{url:"/_next/static/chunks/542.aff58bcde21fcebe.js",revision:"aff58bcde21fcebe"},{url:"/_next/static/chunks/61-143f7360a9128199.js",revision:"143f7360a9128199"},{url:"/_next/static/chunks/821-63f2881b7edccfcc.js",revision:"63f2881b7edccfcc"},{url:"/_next/static/chunks/894.884bc9854a9cfe85.js",revision:"884bc9854a9cfe85"},{url:"/_next/static/chunks/910-efedfa2b79469cd5.js",revision:"efedfa2b79469cd5"},{url:"/_next/static/chunks/991-a04c60806e2a5bc2.js",revision:"a04c60806e2a5bc2"},{url:"/_next/static/chunks/framework-a87821de553db91d.js",revision:"a87821de553db91d"},{url:"/_next/static/chunks/main-21fbc9a5ad2df4ee.js",revision:"21fbc9a5ad2df4ee"},{url:"/_next/static/chunks/pages/404-74a321d3345714e6.js",revision:"74a321d3345714e6"},{url:"/_next/static/chunks/pages/_app-ea8f51670083e2f0.js",revision:"ea8f51670083e2f0"},{url:"/_next/static/chunks/pages/_error-815e492bede44f3e.js",revision:"815e492bede44f3e"},{url:"/_next/static/chunks/pages/cartoons-63769e0e4ce8e9b9.js",revision:"63769e0e4ce8e9b9"},{url:"/_next/static/chunks/pages/favourites-dc32126fb1775c6a.js",revision:"dc32126fb1775c6a"},{url:"/_next/static/chunks/pages/film/%5Bid%5D-45a22eadb831ff2d.js",revision:"45a22eadb831ff2d"},{url:"/_next/static/chunks/pages/films-e9b747c9e3a115ff.js",revision:"e9b747c9e3a115ff"},{url:"/_next/static/chunks/pages/index-a420540d96a29ec6.js",revision:"a420540d96a29ec6"},{url:"/_next/static/chunks/pages/login-b9baa859b3da5271.js",revision:"b9baa859b3da5271"},{url:"/_next/static/chunks/pages/name/%5Bid%5D-8ff9d4aca5802f7b.js",revision:"8ff9d4aca5802f7b"},{url:"/_next/static/chunks/pages/register-829822fc1c8561c7.js",revision:"829822fc1c8561c7"},{url:"/_next/static/chunks/pages/room/%5Bid%5D-e1908d649273067d.js",revision:"e1908d649273067d"},{url:"/_next/static/chunks/pages/search/%5Bid%5D-2b5f9b075f959080.js",revision:"2b5f9b075f959080"},{url:"/_next/static/chunks/pages/series-69da5cdcbac5b1f4.js",revision:"69da5cdcbac5b1f4"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-08d05dc8a1f47353.js",revision:"08d05dc8a1f47353"},{url:"/_next/static/css/0c24d2c5563bb4d1.css",revision:"0c24d2c5563bb4d1"},{url:"/_next/static/css/15e4c5310504cd54.css",revision:"15e4c5310504cd54"},{url:"/_next/static/css/22d5d7579d114f75.css",revision:"22d5d7579d114f75"},{url:"/_next/static/css/390579dc1b8fbe64.css",revision:"390579dc1b8fbe64"},{url:"/_next/static/css/62daf9081586b583.css",revision:"62daf9081586b583"},{url:"/_next/static/css/81e102bbba6f5df3.css",revision:"81e102bbba6f5df3"},{url:"/_next/static/css/ac1370f24fa37b0c.css",revision:"ac1370f24fa37b0c"},{url:"/_next/static/css/b4bf9abad4c08e27.css",revision:"b4bf9abad4c08e27"},{url:"/_next/static/css/bda97fb7cc7e3a91.css",revision:"bda97fb7cc7e3a91"},{url:"/_next/static/css/bec5ff6efd5abea0.css",revision:"bec5ff6efd5abea0"},{url:"/_next/static/css/ea0a40089e461740.css",revision:"ea0a40089e461740"},{url:"/_next/static/media/lg.955a4bcf.woff2",revision:"955a4bcf"},{url:"/_next/static/media/lg.a5ca0178.svg",revision:"a5ca0178"},{url:"/_next/static/media/lg.c950f0b5.woff",revision:"c950f0b5"},{url:"/_next/static/media/lg.dc565ab5.ttf",revision:"dc565ab5"},{url:"/_next/static/media/loading.49ca460c.gif",revision:"49ca460c"},{url:"/bg1.jpg",revision:"d75fc0ac993186a6d2f772fb906eabeb"},{url:"/favicon.ico",revision:"0b5d739697aa36f8f02fcc05cf17b2bd"},{url:"/fonts/FiraSans-Bold.woff2",revision:"07b9daf36cfedbb07b47c1eed9921763"},{url:"/fonts/FiraSans-ExtraBold.woff2",revision:"84e01280cf056a4198d6c14b2f8db407"},{url:"/fonts/FiraSans-Light.woff2",revision:"c2ece2ce8559a936160dd98c0747c9b2"},{url:"/fonts/FiraSans-Medium.woff2",revision:"3392354b6a7a83ac879b2dd4e258e7e1"},{url:"/fonts/FiraSans-Regular.woff2",revision:"44af09a058cd81661fc93376ee5d8f34"},{url:"/fonts/FiraSans-SemiBold.woff2",revision:"0b6a215f82eda2e94a2734e24a3924df"},{url:"/icons/icon-128x128.png",revision:"56d4125eeb03dcd0ec45a8b7c8448123"},{url:"/icons/icon-144x144.png",revision:"c616d2acddf9ad9c887616fee28998d9"},{url:"/icons/icon-152x152.png",revision:"4166a839dff25fa67fbc91461cc0d0f9"},{url:"/icons/icon-192x192.png",revision:"9dd9391d29f0a94f6a5014952b5bc94a"},{url:"/icons/icon-384x384.png",revision:"14e26ef81ef4b67665ea9af23e6fadcb"},{url:"/icons/icon-512x512.png",revision:"6302c46f65bb9ae7d18298a8e9815039"},{url:"/icons/icon-72x72.png",revision:"99f9d2024ae87162655c5120fc41cc74"},{url:"/icons/icon-96x96.png",revision:"689d68e5664c97d5bacb763a8c3468a3"},{url:"/kinobd.txt",revision:"b6d00bafd9a20676f430e8f3adbf3f41"},{url:"/logo.svg",revision:"aa4cf4f36c21cbbf39d36c4ef9455f78"},{url:"/manifest.json",revision:"127550a4d147e1e46235023b66936c18"},{url:"/player.js",revision:"ad82e20958c8ebd61c0e030040a18748"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
