+++ src/components/screens/Home/index.ts
@@ -1,0 +1,5 @@
+import {Hero} from './components/Hero/Hero'
+import {NewFilms} from './components/NewMovies/NewFilms'
+import {NewSeries} from './components/NewMovies/NewSeries'
+
+export {Hero, NewFilms, NewSeries}
