+++ src/pages/cartoons.tsx
@@ -1,5 +1,5 @@
 import { GetStaticProps, NextPage } from 'next';
-import { Cartoons } from '@/components/screens/Cartoons/Cartoons';
+import { Cartoons } from '@/screens/Cartoons/Cartoons';
 import { getCartoons } from '@/services/KinomoreService';
 import { initStore } from '@/store/store';
 import { Layout } from '@/components/Layout/Layout';
