+++ src/pages/favourites.tsx
@@ -1,5 +1,5 @@
 import { Layout } from '@/components/Layout/Layout';
-import { Favourites } from '@/components/screens/Favourites/Favourites';
+import { Favourites } from '@/screens/Favourites/Favourites';
 import { NextPage } from 'next';
 
 const FavouritesPage: NextPage = () => {
