+++ src/pages/login.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next';
 import { Layout } from '@/components/Layout/Layout';
-import { SignIn } from '@/components/screens/SignIn/SignIn';
+import { SignIn } from '@/screens/SignIn/SignIn';
 
 const LoginPage: NextPage = () => {
 	return (
