+++ src/pages/register.tsx
@@ -1,5 +1,5 @@
 import { Layout } from '@/components/Layout/Layout';
-import { SignUp } from '@/components/screens/SignUp/SignUp';
+import { SignUp } from '@/screens/SignUp/SignUp';
 import { NextPage } from 'next';
 
 const RegisterPage: NextPage = () => {
