+++ src/pages/room/[id].tsx
@@ -1,7 +1,7 @@
 import { GetServerSideProps, NextPage } from 'next';
 import { getFilmById } from '@/services/KinomoreService';
 import { initStore } from '@/store/store';
-import { Room } from '@/components/screens/Room/Room';
+import { Room } from '@/screens/Room/Room';
 
 const RoomPage: NextPage = () => {
 	return (
