+++ src/pages/search/[id].tsx
@@ -1,5 +1,5 @@
 import { GetServerSideProps, NextPage } from 'next';
-import { SearchResults } from '@/components/screens/SearchResults/SearchResults';
+import { SearchResults } from '@/screens/SearchResults/SearchResults';
 import { getFilmByName } from '@/services/KinomoreService';
 import { initStore } from '@/store/store';
 import { Layout } from '@/components/Layout/Layout';
